UPDATE: We can now Create and Delete Events | :white_check_mark:
Delete/Destroy also delete associated tags and position but only if the tags and position only had that event left.

Tag to lowercase only, fix view to not include keys because they can be different on database setup as they are being generated automatically.

Resetted the database after bunch of testing.